Honor shutdown timeout in STDIO transport
